Grindstones Restore Max Integrity
Exactly as it says in the title.
A grindstone can now return a weapon to its initial state of sharpness. No longer do you have to toss aside a katar after a small bit of use, or a unique weapon because it's effectively worthless after some repairs and sharpening.

Includes a tutorialization of this in weapon tips, too.