Ducal Standard Bearer
An entirely new subclass for the MAAs, limited to one slot, focused on staying at the sergeant's side. They wield a pseudo-sapient weapon for this purpose, allowing them to rally garrison members.

They can apply a moodbuff, call retinue up to 75 tiles away and have a rather powerful implement for this purpose. There were more abilities planned for this, but figure I'd toss it up just so folks can see it.

The standard itself is effectively a one-force better boar spear, with some funny intents, much higher durability, the verbs and some other bonuses. The wielder, standard bearer or not, as long as they're an MAA, is given +3 luck, hard dismember and no damage slowdown. It does not provide these if not in hand, and it can't be stored aside from using a GWS. Only those with the standard bearer trait (the standard bearer!!!!) is given the rally and plant verb(moodbuff).

To use rally or plant, there's an eight second windup and it takes a good chunk of stamina.

As to the role itself, it's given middling stats and skills, while carried entirely by the in town guard bonus, the banner's buffs itself to the user, and their crit resist trait. Given they have no armour training whatsoever, or dodge expert.